Remove unused interface
